+++ built_value_navigation_saver/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.2] - 09.05.2020
+
+* Navigation 2.0 migration
+* Dependencies upgrade
+
 ## [0.2.1] - 22.11.2019
 
 * Read.me update