+++ shared_pref_navigation_saver/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.3.2] - 09.05.2020
+
+* Dependencies upgrade
+
+## [0.3.1] - 09.05.2020
+
+* Navigation 2.0 migration
+* Dependencies upgrade
+
 ## [0.2.1] - 25.11.2019
 
 * Read.me update